Rename cgclosure to cginfo
