+++ src/cginfo.h
