+++ src/sfpcheck.cpp
@@ -20,7 +20,7 @@
 #include <llvm/Support/raw_ostream.h>
 
 #include "allocators.h"
-#include "cgclosure.h"
+#include "cginfo.h"
 #include "exceptions.h"
 #include "lannotate.h"
