+++ src/ueacheck.cpp
@@ -24,7 +24,7 @@
 #include <llvm/Support/raw_ostream.h>
 
 #include "allocators.h"
-#include "cgclosure.h"
+#include "cginfo.h"
 
 using namespace llvm;
